+++ src/app/(home)/(onboarding)/layout.tsx
@@ -1,0 +1,5 @@
+// src/app/(onboarding)/layout.tsx
+export default function OnboardingLayout({ children }: { children: React.ReactNode }) {
+    return <>{children}</>;
+  }
+
