+++ src/app/(home)/layout.tsx
@@ -11,9 +11,9 @@ export default function HomeLayout({ children }: PropsWithChildren) {
     <div className="flex min-h-screen">
       <Sidebar />
 
-      <div className="w-full">
+      <div className="w-full flex flex-col flex-1">
         <Header />
-        <main className="isolate mx-auto w-full max-w-screen-2xl overflow-hidden p-4 md:p-6 2xl:p-10">
+        <main className="isolate mx-auto w-full max-w-screen-2xl overflow-hidden p-4 md:p-6 2xl:p-10 flex-1">
           {children}
         </main>
       </div>
