+++ src/app/api/create-checkout-session/route.ts
@@ -2,7 +2,7 @@
 
 import { NextRequest, NextResponse } from 'next/server';
 import Stripe from 'stripe';
-import { auth } from '../../../../auth'; // adjust this path as needed
+import { auth } from '../../../auth'; // adjust this path as needed
 import { prisma } from '../../../../prisma';
 
 const stripe = new Stripe(process.env.STRIPE_SECRET_KEY!, {
